+++ src/devtools.d.ts
@@ -1,0 +1,3 @@
+import { Enhancer } from './index';
+
+export const reduxDevToolsExt: <S, A>() => Enhancer<S, A>;