+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Changed
+- Hacky/dirty Redux DevTools Extension support
 - Fix type definition
 
 ## [0.1.0] - 2018-11-07